backend/requirements.txt
fastapi==0.85.0
loguru==0.6.0
numpy==1.23.2
psycopg2==2.9.3
pydantic==1.9.2
python-dotenv==0.21.0
spacy==3.4.1
SQLAlchemy==1.4.41
torch==1.12.1
transformers==4.21.3
uvicorn==0.18.3
websockets==10.3
